refactor(ffi-bridge): split vault.rs into vault/ directory module
Pure refactor: 989-line vault.rs becomes a 5-file directory module per
issue #36. Public surface (crate::vault::{open_vault_with_password,
open_vault_with_recovery, OpenVaultOutput, OpenVaultManifest,
BlockSummary}) is unchanged; bridge-internal pub(crate) accessors stay
in the same scope.

- vault/mod.rs (54)        — module docs + re-exports
- vault/inner.rs (67)      — OpenVaultManifestInner + BlockSummary
- vault/manifest.rs (332)  — OpenVaultManifest + accessors +
                              ReplaceManifestError + block_entry_to_summary
- vault/orchestration.rs (126) — OpenVaultOutput + open_vault_with_password +
                                  open_vault_with_recovery +
                                  split_core_open_vault
- vault/tests.rs (462)     — all integration tests against
                              golden_vault_NNN fixtures

Test counts unchanged: 599 passed + 9 ignored.
